Added callbackManager and renamed all member variables with m_ prefix

+++ C_Plus_Plus_Game/HealthUi.h
@@ -46,7 +46,7 @@ public:
 		m_height = height;
 	}
 
-	void updateUIOnDamage(const float& initialHealth, const  float& currentHealth)
+	void updateUIOnDamage(const int& initialHealth, const  int& currentHealth)
 	{
 		m_startTime = *m_state->getPausableClock();
 		//reduces fill size, so that the red background appears
